Rename SubscriptionAssignModal to SubscriptionAssign
